Refatorado classes para teste unitário em ProductController

+++ src/main/java/com/mercadolibre/demo/service/ProductService.java
@@ -19,7 +19,7 @@ public class ProductService {
 	}
 
 	public Product save(ProductDTO dto) {
-		 Product product = convertProductToDTO(dto);
+		Product product = convertProductToDTO(dto);
 		return productRepository.save(product);
 	}
 
@@ -42,8 +42,13 @@ public class ProductService {
 		}
 	}
 
-	public void delete(Long id) {
-		productRepository.deleteById(id);
+	public void delete(Long id) throws Exception {
+		Optional<Product> existProduct = findById(id);
+		if(existProduct.isPresent()) {
+			productRepository.deleteById(id);
+		} else {
+			throw new Exception("Id não cadastrado");
+		}
 	}
 
 	public Product convertProductToDTO(ProductDTO dto) {
